create separate reducer file
