+++ src/app/redux/GlobalReducer.js
@@ -32,10 +32,6 @@ const SHOW_DIALOG = 'global/SHOW_DIALOG';
 const HIDE_DIALOG = 'global/HIDE_DIALOG';
 const ADD_ACTIVE_WITNESS_VOTE = 'global/ADD_ACTIVE_WITNESS_VOTE';
 const REMOVE_ACTIVE_WITNESS_VOTE = 'global/REMOVE_ACTIVE_WITNESS_VOTE';
-const ADD_ACTIVE_PROPOSAL_VOTE = 'global/ADD_ACTIVE_PROPOSAL_VOTE';
-const REMOVE_ACTIVE_PROPOSAL_VOTE = 'global/REMOVE_ACTIVE_PROPOSAL_VOTE';
-const RECEIVE_LIST_PROPOSALS = 'global/RECEIVE_LIST_PROPOSALS';
-const RECEIVE_LIST_VOTER_PROPOSALS = 'global/RECEIVE_LIST_VOTER_PROPOSALS';
 // Saga-related:
 export const GET_STATE = 'global/GET_STATE';
 
@@ -326,32 +322,6 @@ export default function reducer(state = defaultState, action = {}) {
             );
         }
 
-        case ADD_ACTIVE_PROPOSAL_VOTE: {
-            return state.update(
-                `transaction_proposal_vote_active_${payload.voter}`,
-                List(),
-                l => l.push(payload.proposal_ids[0])
-            );
-        }
-
-        case REMOVE_ACTIVE_PROPOSAL_VOTE: {
-            return state.update(
-                `transaction_proposal_vote_active_${payload.voter}`,
-                List(),
-                l => l.delete(l.indexOf(payload.proposal_ids[0]))
-            );
-        }
-
-        case RECEIVE_LIST_PROPOSALS: {
-            const new_state = fromJS(payload);
-            return state.merge(new_state);
-        }
-
-        case RECEIVE_LIST_VOTER_PROPOSALS: {
-            const new_state = fromJS(payload);
-            return state.merge(new_state);
-        }
-
         default:
             return state;
     }
@@ -467,26 +437,6 @@ export const addActiveWitnessVote = payload => ({
 
 export const removeActiveWitnessVote = payload => ({
     type: REMOVE_ACTIVE_WITNESS_VOTE,
-    payload,
-});
-
-export const addActiveProposalVote = payload => ({
-    type: ADD_ACTIVE_PROPOSAL_VOTE,
-    payload,
-});
-
-export const removeActiveProposalVote = payload => ({
-    type: REMOVE_ACTIVE_PROPOSAL_VOTE,
-    payload,
-});
-
-export const receiveListProposals = payload => ({
-    type: RECEIVE_LIST_PROPOSALS,
-    payload,
-});
-
-export const receiveListVoterProposals = payload => ({
-    type: RECEIVE_LIST_VOTER_PROPOSALS,
     payload,
 });
